Make minor changes (reference correct fields and change order of setting fields) in Elevator subsystem. Move some classes out of elevator (which required changing imports on other classes).
