+++ src/main/java/frc/robot/subsystems/MotionProfile.java
@@ -11,6 +11,7 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj.command.Subsystem;
+import frc.robot.subsystems.elevator.Elevator;
 import frc.robot.utils.RobotTime;
 
 /**
